refactor: ReservationAcceptanceTest 클래스를 시나리오 별로 분리
